+++ click/types.py
@@ -495,9 +495,9 @@ class Path(ParamType):
                       input data Click deals with.
     :param converter: optionally a callable to convert the otherwise final
                       result of the parameter.  For example passing
-                      `converter=pathlib.Path` will result in a `pathlib.Path`
-                      object being passed to the command rather than the normal
-                      `str`.
+                      ``converter=pathlib.Path`` will result in a
+                      :class:`pathlib.Path` object being passed to the command
+                      rather than the normal :class:`str`.
     """
     envvar_list_splitter = os.path.pathsep